+++ packages/build-plugin-component/src/constants.js
@@ -5,4 +5,7 @@ module.exports = {
   WECHAT_MINIPROGRAM: 'wechat-miniprogram',
   BYTEDANCE: 'bytedance-microapp',
   NODE: 'node',
+  KRAKEN: 'kraken',
+  RUNTIME: 'runtime',
+  COMPILE: 'compile',
 };