+++ packages/ice-npm-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.1.1
+
+- [chore] better error message
+
+## 2.1.0
+- [feat] add `getVersions` and `getSatisfiesVersions` method
+
 ## 1.4.1
 
 - [chore] remove cacheData for server use